+++ app/src/main/java/uk/org/openseizuredetector/SdDataSourcePebble.java
@@ -24,7 +24,9 @@
 package uk.org.openseizuredetector;
 
 import android.content.Context;
+import android.content.Intent;
 import android.content.SharedPreferences;
+import android.net.Uri;
 import android.os.Handler;
 import android.os.Looper;
 import android.preference.PreferenceManager;
@@ -36,6 +38,11 @@ import com.getpebble.android.kit.Constants;
 import com.getpebble.android.kit.PebbleKit;
 import com.getpebble.android.kit.util.PebbleDictionary;
 
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
 import java.nio.IntBuffer;
@@ -50,13 +57,13 @@ import java.util.UUID;
  * network data source.
  */
 public class SdDataSourcePebble extends SdDataSource {
+    private Handler mHandler = new Handler();
     private Timer mSettingsTimer;
     private Timer mStatusTimer;
     private Time mPebbleStatusTime;
     private boolean mPebbleAppRunningCheck = false;
-    private int mDataPeriod = 5;    // Period at which data is sent from watch to phone (sec)
     private int mAppRestartTimeout = 10;  // Timeout before re-starting watch app (sec) if we have not received
-                                           // data after mDataPeriod
+                                           // data after mDataUpdatePeriod
     //private Looper mServiceLooper;
     private int mFaultTimerPeriod = 30;  // Fault Timer Period in sec
     private PebbleKit.PebbleDataReceiver msgDataHandler = null;
@@ -99,9 +106,26 @@ public class SdDataSourcePebble extends SdDataSource {
     private int DATA_TYPE_RESULTS = 1;   // Analysis Results
     private int DATA_TYPE_SETTINGS = 2;  // Settings
     private int DATA_TYPE_SPEC = 3;      // FFT Spectrum (or part of a spectrum)
+    private short mDataUpdatePeriod;
+    private short mMutePeriod;
+    private short mManAlarmPeriod;
+    private short mAlarmFreqMin;
+    private short mAlarmFreqMax;
+    private short mWarnTime;
+    private short mAlarmTime;
+    private short mAlarmThresh;
+    private short mAlarmRatioThresh;
+    private boolean mFallActive;
+    private short mFallThreshMin;
+    private short mFallThreshMax;
+    private short mFallWindow;
+
     public SdDataSourcePebble(Context context, SdDataReceiver sdDataReceiver) {
         super(context,sdDataReceiver);
         mName = "Pebble";
+        // Set default settings from XML files (mContext is set by super().
+        PreferenceManager.setDefaultValues(mContext,
+                R.xml.pebble_datasource_prefs, true);
     }
 
 
@@ -125,7 +149,7 @@ public class SdDataSourcePebble extends SdDataSource {
                 public void run() {
                     getPebbleStatus();
                 }
-            }, 0, mDataPeriod * 1000);
+            }, 0, mDataUpdatePeriod * 1000);
         } else {
             Log.v(TAG, "onCreate(): status timer already running.");
         }
@@ -141,7 +165,7 @@ public class SdDataSourcePebble extends SdDataSource {
                 public void run() {
                     getPebbleSdSettings();
                 }
-            }, 0, 1000 * (mDataPeriod + 60));  // ask for settings less frequently than we get data
+            }, 0, 1000 * (mDataUpdatePeriod + 60));  // ask for settings less frequently than we get data
         } else {
             Log.v(TAG, "onCreate(): settings timer already running.");
         }
@@ -200,17 +224,6 @@ public class SdDataSourcePebble extends SdDataSource {
                 toast.show();
             }
 
-            // Parse the DataPeriod setting.
-            try {
-                String dataPeriodStr = SP.getString("DataPeriod", "5");
-                mDataPeriod = Integer.parseInt(dataPeriodStr);
-                Log.v(TAG, "updatePrefs() - mDataPeriod = " + mDataPeriod);
-            } catch (Exception ex) {
-                Log.v(TAG, "updatePrefs() - Problem with DataPeriod preference!");
-                Toast toast = Toast.makeText(mContext, "Problem Parsing DataPeriod Preference", Toast.LENGTH_SHORT);
-                toast.show();
-            }
-
             // Parse the FaultTimer period setting.
             try {
                 String faultTimerPeriodStr = SP.getString("FaultTimerPeriod", "30");
@@ -224,82 +237,59 @@ public class SdDataSourcePebble extends SdDataSource {
 
 
             // Watch Settings
-            PebbleDictionary setDict = new PebbleDictionary();
-            short intVal;
             String prefStr;
 
-            prefStr = SP.getString("DataUpdatePeriod", "5");
-            intVal = (short) Integer.parseInt(prefStr);
-            Log.v(TAG, "updatePrefs() DataUpdatePeriod = " + intVal);
-            setDict.addInt16(KEY_DATA_UPDATE_PERIOD, intVal);
+            prefStr = SP.getString("PebbleUpdatePeriod", "SET_FROM_XML");
+            mDataUpdatePeriod = (short) Integer.parseInt(prefStr);
+            Log.v(TAG, "updatePrefs() DataUpdatePeriod = " + mDataUpdatePeriod);
 
-            prefStr = SP.getString("MutePeriod", "300");
-            intVal = (short) Integer.parseInt(prefStr);
-            Log.v(TAG, "updatePrefs() MutePeriod = " + intVal);
-            setDict.addInt16(KEY_MUTE_PERIOD, intVal);
+            prefStr = SP.getString("MutePeriod", "SET_FROM_XML");
+            mMutePeriod = (short) Integer.parseInt(prefStr);
+            Log.v(TAG, "updatePrefs() MutePeriod = " + mMutePeriod);
 
-            prefStr = SP.getString("ManAlarmPeriod", "30");
-            intVal = (short) Integer.parseInt(prefStr);
-            Log.v(TAG, "updatePrefs() ManAlarmPeriod = " + intVal);
-            setDict.addInt16(KEY_MAN_ALARM_PERIOD, intVal);
+            prefStr = SP.getString("ManAlarmPeriod", "SET_FROM_XML");
+            mManAlarmPeriod = (short) Integer.parseInt(prefStr);
+            Log.v(TAG, "updatePrefs() ManAlarmPeriod = " + mManAlarmPeriod);
 
+            prefStr = SP.getString("AlarmFreqMin","SET_FROM_XML");
+            mAlarmFreqMin = (short) Integer.parseInt(prefStr);
+            Log.v(TAG, "updatePrefs() AlarmFreqMin = " + mAlarmFreqMin);
 
-            prefStr = SP.getString("AlarmFreqMin", "5");
-            intVal = (short) Integer.parseInt(prefStr);
-            Log.v(TAG, "updatePrefs() AlarmFreqMin = " + intVal);
-            setDict.addInt16(KEY_ALARM_FREQ_MIN, intVal);
+            prefStr = SP.getString("AlarmFreqMax", "SET_FROM_XML");
+            mAlarmFreqMax = (short) Integer.parseInt(prefStr);
+            Log.v(TAG, "updatePrefs() AlarmFreqMax = " + mAlarmFreqMax);
 
-            prefStr = SP.getString("AlarmFreqMax", "10");
-            intVal = (short) Integer.parseInt(prefStr);
-            Log.v(TAG, "updatePrefs() AlarmFreqMax = " + intVal);
-            setDict.addUint16(KEY_ALARM_FREQ_MAX, (short) intVal);
+            prefStr = SP.getString("WarnTime", "SET_FROM_XML");
+            mWarnTime = (short) Integer.parseInt(prefStr);
+            Log.v(TAG, "updatePrefs() WarnTime = " + mWarnTime);
 
-            prefStr = SP.getString("WarnTime", "5");
-            intVal = (short) Integer.parseInt(prefStr);
-            Log.v(TAG, "updatePrefs() WarnTime = " + intVal);
-            setDict.addUint16(KEY_WARN_TIME, (short) intVal);
+            prefStr = SP.getString("AlarmTime", "SET_FROM_XML");
+            mAlarmTime = (short) Integer.parseInt(prefStr);
+            Log.v(TAG, "updatePrefs() AlarmTime = " + mAlarmTime);
 
-            prefStr = SP.getString("AlarmTime", "10");
-            intVal = (short) Integer.parseInt(prefStr);
-            Log.v(TAG, "updatePrefs() AlarmTime = " + intVal);
-            setDict.addUint16(KEY_ALARM_TIME, (short) intVal);
+            prefStr = SP.getString("AlarmThresh", "SET_FROM_XML");
+            mAlarmThresh = (short) Integer.parseInt(prefStr);
+            Log.v(TAG, "updatePrefs() AlarmThresh = " + mAlarmThresh);
 
-            prefStr = SP.getString("AlarmThresh", "70");
-            intVal = (short) Integer.parseInt(prefStr);
-            Log.v(TAG, "updatePrefs() AlarmThresh = " + intVal);
-            setDict.addUint16(KEY_ALARM_THRESH, (short) intVal);
+            prefStr = SP.getString("AlarmRatioThresh", "SET_FROM_XML");
+            mAlarmRatioThresh = (short) Integer.parseInt(prefStr);
+            Log.v(TAG, "updatePrefs() AlarmRatioThresh = " + mAlarmRatioThresh);
 
-            prefStr = SP.getString("AlarmRatioThresh", "30");
-            intVal = (short) Integer.parseInt(prefStr);
-            Log.v(TAG, "updatePrefs() AlarmRatioThresh = " + intVal);
-            setDict.addUint16(KEY_ALARM_RATIO_THRESH, (short) intVal);
+            mFallActive = SP.getBoolean("FallActive", false);
+            Log.v(TAG, "updatePrefs() FallActive = " + mFallActive);
 
-            boolean fallActiveBool = SP.getBoolean("FallActive", false);
-            Log.v(TAG, "updatePrefs() FallActive = " + fallActiveBool);
-            if (fallActiveBool)
-                setDict.addUint16(KEY_FALL_ACTIVE, (short) 1);
-            else
-                setDict.addUint16(KEY_FALL_ACTIVE, (short) 0);
+            prefStr = SP.getString("FallThreshMin", "SET_FROM_XML");
+            mFallThreshMin = (short) Integer.parseInt(prefStr);
+            Log.v(TAG, "updatePrefs() FallThreshMin = " + mFallThreshMin);
 
-            prefStr = SP.getString("FallThreshMin", "200");
-            intVal = (short) Integer.parseInt(prefStr);
-            Log.v(TAG, "updatePrefs() FallThreshMin = " + intVal);
-            setDict.addUint16(KEY_FALL_THRESH_MIN, (short) intVal);
+            prefStr = SP.getString("FallThreshMax", "SET_FROM_XML");
+            mFallThreshMax = (short) Integer.parseInt(prefStr);
+            Log.v(TAG, "updatePrefs() FallThreshMax = " + mFallThreshMax);
 
-            prefStr = SP.getString("FallThreshMax", "1200");
-            intVal = (short) Integer.parseInt(prefStr);
-            Log.v(TAG, "updatePrefs() FallThreshMax = " + intVal);
-            setDict.addUint16(KEY_FALL_THRESH_MAX, (short) intVal);
+            prefStr = SP.getString("FallWindow", "SET_FROM_XML");
+            mFallWindow = (short) Integer.parseInt(prefStr);
+            Log.v(TAG, "updatePrefs() FallWindow = " + mFallWindow);
 
-            prefStr = SP.getString("FallWindow", "1500");
-            intVal = (short) Integer.parseInt(prefStr);
-            Log.v(TAG, "updatePrefs() FallWindow = " + intVal);
-            setDict.addUint16(KEY_FALL_WINDOW, (short) intVal);
-
-
-            // Send Watch Settings to Pebble
-            Log.v(TAG, "updatePrefs() - setDict = " + setDict.toJsonString());
-            PebbleKit.sendDataToPebble(mContext, SD_UUID, setDict);
         } catch (Exception ex) {
             Log.v(TAG, "updatePrefs() - Problem parsing preferences!");
             Toast toast = Toast.makeText(mContext, "Problem Parsing Preferences - Something won't work - Please go back to Settings and correct it!", Toast.LENGTH_SHORT);
@@ -322,7 +312,7 @@ public class SdDataSourcePebble extends SdDataSource {
                                     final PebbleDictionary data) {
                 Log.v(TAG, "Received message from Pebble - data type="
                         + data.getUnsignedIntegerAsLong(KEY_DATA_TYPE));
-                // If we ha ve a message, the app must be running
+                // If we have a message, the app must be running
                 Log.v(TAG,"Setting mPebbleAppRunningCheck to true");
                 mPebbleAppRunningCheck = true;
                 PebbleKit.sendAckToPebble(context, transactionId);
@@ -342,13 +332,12 @@ public class SdDataSourcePebble extends SdDataSource {
                     mSdData.alarmPhrase = "Unknown";
                     mSdData.haveData = true;
                     mSdDataReceiver.onSdDataReceived(mSdData);
-                    }
 
 
                     // Read the data that has been sent, and convert it into
                     // an integer array.
                     byte[] byteArr = data.getBytes(KEY_SPEC_DATA);
-                    if ((byteArr!=null) && (byteArr.length!=0)) {
+                    if ((byteArr != null) && (byteArr.length != 0)) {
                         IntBuffer intBuf = ByteBuffer.wrap(byteArr)
                                 .order(ByteOrder.LITTLE_ENDIAN)
                                 .asIntBuffer();
@@ -358,9 +347,9 @@ public class SdDataSourcePebble extends SdDataSource {
                             mSdData.simpleSpec[i] = intArray[i];
                         }
                     } else {
-                        Log.v(TAG,"***** zero length spectrum received - error!!!!");
+                        Log.v(TAG, "***** zero length spectrum received - error!!!!");
                     }
-
+                }
 
                 if (data.getUnsignedIntegerAsLong(KEY_DATA_TYPE)
                         == DATA_TYPE_SETTINGS) {
@@ -379,9 +368,10 @@ public class SdDataSourcePebble extends SdDataSource {
             }
         };
         PebbleKit.registerReceivedDataHandler(mContext, msgDataHandler);
-        // We struggle to connect to pebble time if app is already running, so stop app so we can
-        // re-connect to it.
+        // We struggle to connect to pebble time if app is already running,
+        // so stop app so we can re-connect to it.
         stopWatchApp();
+        startWatchApp();
     }
 
     /**
@@ -416,10 +406,13 @@ public class SdDataSourcePebble extends SdDataSource {
 
 
     /**
-     * Request Pebble App to send us its latest settings.
+     * Send our latest settings to the watch, then request Pebble App to send
+     * us its latest settings so we can check it has been set up correctly..
      * Will be received as a message by the receiveData handler
      */
     public void getPebbleSdSettings() {
+        Log.v(TAG, "getPebbleSdSettings() - sending required settings to pebble");
+        sendPebbleSdSettings();
         Log.v(TAG, "getPebbleSdSettings() - requesting settings from pebble");
         PebbleDictionary data = new PebbleDictionary();
         data.addUint8(KEY_SETTINGS, (byte) 1);
@@ -427,6 +420,101 @@ public class SdDataSourcePebble extends SdDataSource {
                 mContext,
                 SD_UUID,
                 data);
+    }
+
+    /**
+     * Send the pebble watch settings that are stored as class member
+     * variables to the watch.
+     */
+    public void sendPebbleSdSettings() {
+        Log.v(TAG, "sendPebblSdSettings() - preparing settings dictionary..");
+        // Watch Settings
+        final PebbleDictionary setDict = new PebbleDictionary();
+        setDict.addInt16(KEY_DATA_UPDATE_PERIOD, mDataUpdatePeriod);
+        setDict.addInt16(KEY_MUTE_PERIOD, mMutePeriod);
+        setDict.addInt16(KEY_MAN_ALARM_PERIOD, mManAlarmPeriod);
+        setDict.addInt16(KEY_ALARM_FREQ_MIN, mAlarmFreqMin);
+        setDict.addInt16(KEY_ALARM_FREQ_MAX, mAlarmFreqMax);
+        setDict.addUint16(KEY_WARN_TIME, mWarnTime);
+        setDict.addUint16(KEY_ALARM_TIME, mAlarmTime);
+        setDict.addUint16(KEY_ALARM_THRESH, mAlarmThresh);
+        setDict.addUint16(KEY_ALARM_RATIO_THRESH, mAlarmRatioThresh);
+        if (mFallActive)
+            setDict.addUint16(KEY_FALL_ACTIVE, (short) 1);
+        else
+            setDict.addUint16(KEY_FALL_ACTIVE, (short) 0);
+        setDict.addUint16(KEY_FALL_THRESH_MIN, mFallThreshMin);
+        setDict.addUint16(KEY_FALL_THRESH_MAX, mFallThreshMax);
+        setDict.addUint16(KEY_FALL_WINDOW, mFallWindow);
+
+        // Send Watch Settings to Pebble
+        Log.v(TAG, "sendPebbleSdSettings() - setDict = " + setDict.toJsonString());
+        PebbleKit.sendDataToPebble(mContext, SD_UUID, setDict);
+    }
+
+
+    /**
+     * Compares the watch settings retrieved from the watch (stored in mSdData)
+     * to the required settings stored as member variables to this class.
+     *
+     * @return true if they are all the same, or false if there are discrepancies.
+     */
+    public boolean checkWatchSettings() {
+       boolean settingsOk = true;
+        if (mDataUpdatePeriod !=  mSdData.mDataUpdatePeriod) {
+            Log.v(TAG,"checkWatchSettings - mDataUpdatePeriod Wrong");
+            settingsOk = false;
+        }
+        if (mMutePeriod != mSdData.mMutePeriod) {
+            Log.v(TAG,"checkWatchSettings - mMutePeriod Wrong");
+            settingsOk = false;
+        }
+        if (mManAlarmPeriod != mSdData.mManAlarmPeriod) {
+            Log.v(TAG,"checkWatchSettings - mManAlarmPeriod Wrong");
+            settingsOk = false;
+        }
+        if (mAlarmFreqMin != mSdData.alarmFreqMin) {
+            Log.v(TAG,"checkWatchSettings - mAlarmFreqMin Wrong");
+            settingsOk = false;
+        }
+        if (mAlarmFreqMax != mSdData.alarmFreqMax) {
+            Log.v(TAG,"checkWatchSettings - mAlarmFreqMax Wrong");
+            settingsOk = false;
+        }
+        if (mWarnTime != mSdData.warnTime) {
+            Log.v(TAG,"checkWatchSettings - mWarnTime Wrong");
+            settingsOk = false;
+        }
+        if (mAlarmTime != mSdData.alarmTime) {
+            Log.v(TAG,"checkWatchSettings - mAlarmTime Wrong");
+            settingsOk = false;
+        }
+        if (mAlarmThresh != mSdData.alarmThresh) {
+            Log.v(TAG,"checkWatchSettings - mAlarmThresh Wrong");
+            settingsOk = false;
+        }
+        if (mAlarmRatioThresh != mSdData.alarmRatioThresh) {
+            Log.v(TAG,"checkWatchSettings - mAlarmRatioThresh Wrong");
+            settingsOk = false;
+        }
+        if (mFallActive != mSdData.mFallActive) {
+            Log.v(TAG,"checkWatchSettings - mAlarmFreqMin Wrong");
+            settingsOk = false;
+        }
+        if (mFallThreshMin != mSdData.mFallThreshMin) {
+            Log.v(TAG,"checkWatchSettings - mFallThreshMin Wrong");
+            settingsOk = false;
+        }
+        if (mFallThreshMax != mSdData.mFallThreshMax) {
+            Log.v(TAG,"checkWatchSettings - mFallThreshMax Wrong");
+            settingsOk = false;
+        }
+        if (mFallWindow != mSdData.mFallWindow) {
+            Log.v(TAG,"checkWatchSettings - mFallWindow Wrong");
+            settingsOk = false;
+        }
+
+        return settingsOk;
     }
 
     /**
@@ -464,7 +552,7 @@ public class SdDataSourcePebble extends SdDataSource {
         // the app is not talking to us
         // mPebbleAppRunningCheck is set to true in the receiveData handler.
         if (!mPebbleAppRunningCheck &&
-                (tdiff > (mDataPeriod+mAppRestartTimeout) * 1000)) {
+                (tdiff > (mDataUpdatePeriod+mAppRestartTimeout) * 1000)) {
             Log.v(TAG, "getPebbleStatus() - tdiff = " + tdiff);
             mSdData.pebbleAppRunning = false;
             Log.v(TAG, "getPebbleStatus() - Pebble App Not Running - Attempting to Re-Start");
@@ -472,7 +560,7 @@ public class SdDataSourcePebble extends SdDataSource {
             //mPebbleStatusTime = tnow;  // set status time to now so we do not re-start app repeatedly.
             getPebbleSdSettings();
             // Only make audible warning beep if we have not received data for more than mFaultTimerPeriod seconds.
-            if (tdiff > (mDataPeriod+mFaultTimerPeriod) * 1000) {
+            if (tdiff > (mDataUpdatePeriod+mFaultTimerPeriod) * 1000) {
                 mSdDataReceiver.onSdDataFault(mSdData);
             } else {
                 Log.v(TAG, "getPebbleStatus() - Waiting for mFaultTimerPeriod before issuing audible warning...");
@@ -494,6 +582,39 @@ public class SdDataSourcePebble extends SdDataSource {
             getPebbleData();
         }
     }
+
+    /**
+     * Install the wach app that is bundled in the 'assets' folder of this
+     * phone app.
+     * from https://github.com/pebble-examples/pebblekit-android-example/blob/master/android/Eclipse/src/com/getpebble/pebblekitexample/MainActivity.java#L148
+     */
+    @Override
+    public void installWatchApp() {
+        Log.v(TAG, "SdDataSourcePebble.installWatchApp()");
+        final String WATCHAPP_FILENAME = "pebble_sd.pbw";
+
+        try {
+            // Read .pbw from assets/
+            Intent intent = new Intent(Intent.ACTION_VIEW);
+            File file = new File(mContext.getExternalFilesDir(null), WATCHAPP_FILENAME);
+            InputStream is = mContext.getResources().getAssets().open(WATCHAPP_FILENAME);
+            OutputStream os = new FileOutputStream(file);
+            byte[] pbw = new byte[is.available()];
+            is.read(pbw);
+            os.write(pbw);
+            is.close();
+            os.close();
+
+            // Install via Pebble Android app
+            intent.setDataAndType(Uri.fromFile(file), "application/pbw");
+            intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+            mContext.startActivity(intent);
+        } catch (IOException e) {
+            Toast.makeText(mContext, "App install failed: " + e.getLocalizedMessage(), Toast.LENGTH_LONG).show();
+        }
+
+    }
+